requirements.txt
requests
beautifulsoup4
selenium
pymongo
youtube-transcript-api
webdriver-manager
python-dotenv
redis